requirements.txt
streamlit==1.19.0
torch==1.10.0
torchvision==0.11.1
Pillow==8.4.0
numpy==1.21.4
opencv-python==4.5.4.60
pytesseract==0.3.8
